requirements.txt
appnope==0.1.0
APScheduler==3.5.3
attrs==19.1.0
backcall==0.1.0
backports.csv==1.0.7
bleach==2.1.4
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-static==0.0.6
Django==2.0.6
django-apscheduler==0.2.13
django-crontab==0.7.1
django-extensions==2.1.0
django-import-export==1.2.0
django-storages==1.6.6
entrypoints==0.2.3
et-xmlfile==1.0.1
future==0.16.0
gunicorn==19.8.1
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
jdcal==1.4.1
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
line-bot-sdk==1.8.0
livereload==2.5.1
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.14.5
odfpy==1.4.0
openpyxl==3.0.0
pandas==0.23.3
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
pip-autoremove==0.9.1
prometheus-client==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
PyMySQL==0.9.3
pyreadline==2.1
pyrsistent==0.15.3
python-dateutil==2.7.3
pytz==2018.4
PyYAML==5.1.2
pyzmq==17.1.2
qtconsole==4.5.1
requests==2.18.4
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.3.7
sqlparse==0.3.0
static3==0.7.0
tablib==0.13.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1
tqdm==4.32.2
traitlets==4.3.2
tzlocal==1.5.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.0
xlrd==1.2.0
xlwt==1.3.0
